Use size_t instead of int, use more specific format string

+++ ext/pdo_dblib/dblib_stmt.c
@@ -330,14 +330,14 @@ static int pdo_dblib_stmt_get_col(pdo_stmt_t *stmt, int colno, char **ptr,
 				}
 				case SQLDATETIME:
 				case SQLDATETIM4: {
-					int dl;
+					size_t dl;
 					DBDATEREC di;
 					DBDATEREC dt;
 
 					dbconvert(H->link, coltype, data, -1, SQLDATETIME, (LPBYTE) &dt, -1);
 					dbdatecrack(H->link, &di, (DBDATETIME *) &dt);
 
-					dl = spprintf(&tmp_data, 20, "%d-%02d-%02d %02d:%02d:%02d",
+					dl = spprintf(&tmp_data, 20, "%04d-%02d-%02d %02d:%02d:%02d",
 #if defined(PHP_DBLIB_IS_MSSQL) || defined(MSDBLIB)
 							di.year,     di.month,       di.day,        di.hour,     di.minute,     di.second
 #else
